+++ src/main/java/org/stellar/sdk/responses/PredicateDeserializer.java
@@ -3,6 +3,10 @@ package org.stellar.sdk.responses;
 import com.google.common.collect.Lists;
 import com.google.gson.*;
 import org.stellar.sdk.Predicate;
+import org.stellar.sdk.xdr.Duration;
+import org.stellar.sdk.xdr.Int64;
+import org.stellar.sdk.xdr.TimePoint;
+import org.stellar.sdk.xdr.Uint64;
 import org.threeten.bp.Instant;
 
 import java.lang.reflect.Type;
@@ -39,15 +43,14 @@ public class PredicateDeserializer implements JsonDeserializer<Predicate> {
     // backwards compatible for abs before, uses the newer abs_before_epoch if available from server
     // otherwise if abs_before_epoch is absent, it will fall back to original attempt to parse abs_before date string
     if (obj.has("abs_before_epoch")) {
-      return new Predicate.AbsBefore(obj.get("abs_before_epoch").getAsLong());
+      return new Predicate.AbsBefore(new TimePoint(new Uint64(obj.get("abs_before_epoch").getAsLong())));
     } else if (obj.has("abs_before")) {
       String formattedDate = obj.get("abs_before").getAsString();
-      return new Predicate.AbsBefore(Instant.parse(formattedDate).getEpochSecond());
+      return new Predicate.AbsBefore(new TimePoint(new Uint64(Instant.parse(formattedDate).getEpochSecond())));
     }
 
     if (obj.has("rel_before")) {
-      Long relBefore = obj.get("rel_before").getAsLong();
-      return new Predicate.RelBefore(relBefore);
+      return new Predicate.RelBefore(new Duration(new Int64(obj.get("rel_before").getAsLong())));
     }
 
     throw new IllegalArgumentException("Unsupported predicate: "+json.toString());